Move device auth flow code into new go file
